pyproject.toml
[tool.poetry]
name = "cv_pipeliner"
version = "0.20.0"
description = ""
authors = [
    "bobokvsky <alexander.kozlovsky.m@gmail.com>",
]

[tool.poetry.dependencies]
python = ">=3.8, <3.12"
opencv-python = ">=4.5.1.48"
numpy = ">=1.19.5"
fsspec = ">=2021.4.0"
pydantic = ">=1.9.2"
imutils = ">=0.5.4"
imageio = ">=2.17.0"
joblib = ">=1.0.1"
pathy = "^0.10.2"
tqdm = ">=4.60.0"
filterpy = ">=1.4.5"
matplotlib = ">=3.4.1"
requests = ">=2.24.0"
scipy = ">=1.6.2"
pandas = ">=1.2.4"
Pillow = ">=8.0.0"
scikit_learn = ">=0.24.2"
lxml = ">=4.6.3"
dataframe-image = ">=0.1.1"
scikit-image = ">=0.18.3"


[tool.poetry.group.dev.dependencies]
pytest = "*"
flake8 = "*"
isort = "*"
black = "*"


[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
